+++ react/package.json
@@ -26,6 +26,7 @@
     "css-loader": "^2.1.0",
     "happypack": "^5.0.1",
     "html-loader": "^0.5.5",
+    "html-webpack-harddisk-plugin": "^1.0.1",
     "html-webpack-plugin": "^3.2.0",
     "mini-css-extract-plugin": "^0.5.0",
     "mobx": "^5.9.0",
