+++ react/src/app/society_bureau/styles/header.scss
